Added table and report details.

+++ src/Queue.php
@@ -6,81 +6,80 @@ defined( 'ABSPATH' ) || exit;
 
 class Queue {
 
-	public static function get_available_types() {
-		return array(
-			'quarterly' => _x( 'Quarterly', 'oss', 'oss-woocommerce' ),
-			'yearly'    => _x( 'Yearly', 'oss', 'oss-woocommerce' ),
-			'monthly'   => _x( 'Monthly', 'oss', 'oss-woocommerce' )
-		);
-	}
-
-	public static function get_type_title( $type ) {
-		$types = self::get_available_types();
-
-		return array_key_exists( $type, $types ) ? $types[ $type ] : '';
-	}
-
-	public static function start( $type = 'quarterly', $date = null ) {
-		$types = self::get_available_types();
+	public static function start( $type = 'quarterly', $date = null, $end_date = null ) {
+		$types = Package::get_available_types();
 
 		if ( ! array_key_exists( $type, $types ) ) {
 			return false;
 		}
 
-		$args       = self::get_timeframe( $type, $date );
+		$args       = self::get_timeframe( $type, $date, $end_date );
 		$generator  = new AsyncReportGenerator( $type, $args );
 		$queue_args = $generator->get_args();
 		$queue      = self::get_queue();
 
-		self::cancel( $type, $args );
+		self::cancel( $generator->get_id() );
 
-		Package::log( sprintf( 'Starting new %1$s report for %2$s - %3$s', self::get_type_title( $type ), $queue_args['start'], $queue_args['end'] ) );
-		Package::extended_log( sprintf( 'Default report arguments: %s', wc_print_r( $queue_args, true ) ) );
+		$report = $generator->start();
 
-		$queue->schedule_single(
-			time() + 10,
-			'oss_woocommerce_' . $type,
-			array( 'args' => $queue_args ),
-			'oss_woocommerce'
-		);
+		if ( is_a( $report, '\Vendidero\OneStopShop\Report' ) && $report->exists() ) {
+			$reports_available = Queue::get_report_ids();
 
-		$running = self::get_reports_running();
-		$running[ $type ] = array( 'id' => $generator->get_id(), 'status' => 'pending' );
+			if ( ! in_array( $report->get_id(), $reports_available[ $type ] ) ) {
+				array_unshift( $reports_available[ $type ], $report->get_id() );
+				update_option( 'oss_woocommerce_reports', $reports_available );
+			}
 
-		update_option( 'oss_woocommerce_reports_running', $running );
+			Package::log( sprintf( 'Starting new %1$s', $report->get_title() ) );
+			Package::extended_log( sprintf( 'Default report arguments: %s', wc_print_r( $queue_args, true ) ) );
 
-		return true;
+			$queue->schedule_single(
+				time() + 10,
+				'oss_woocommerce_' . $generator->get_id(),
+				array( 'args' => $queue_args ),
+				'oss_woocommerce'
+			);
+
+			$running = self::get_reports_running();
+			$running[ $generator->get_id() ] = 'pending';
+
+			update_option( 'oss_woocommerce_reports_running', $running );
+
+			return $generator->get_id();
+		}
+
+		return false;
 	}
 
-	public static function cancel( $type, $args ) {
-		$generator = new AsyncReportGenerator( $type, $args );
+	public static function cancel( $id ) {
+		$data      = Package::get_report_data( $id );
+		$generator = new AsyncReportGenerator( $data['type'], $data );
 		$queue     = self::get_queue();
+		$running   = self::get_reports_running();
 
-		$generator->reset();
+		$generator->delete();
 
-		$running = self::get_reports_running();
-
-		if ( array_key_exists( $type, $running ) ) {
-			unset( $running[ $type ] );
-			Package::log( sprintf( 'Cancelled %s report', self::get_type_title( $type ) ) );
+		if ( array_key_exists( $id, $running ) ) {
+			unset( $running[ $id ] );
+			Package::log( sprintf( 'Cancelled %s', Package::get_report_title( $id ) ) );
 			update_option( 'oss_woocommerce_reports_running', $running );
 		}
 
 		/**
 		 * Cancel outstanding events and queue new.
 		 */
-		$queue->cancel_all( 'oss_woocommerce_' . $type );
+		$queue->cancel_all( 'oss_woocommerce_' . $id );
 	}
 
 	public static function get_queue() {
 		return WC()->queue();
 	}
 
-	public static function is_running( $type ) {
+	public static function is_running( $id ) {
 		$running = self::get_reports_running();
 
-		if ( array_key_exists( $type, $running ) && 'pending' === $running[ $type ]['status'] ) {
-			if ( self::get_queue()->get_next( 'oss_woocommerce_' . $type ) ) {
+		if ( array_key_exists( $id, $running ) && 'pending' === $running[ $id ] ) {
+			if ( self::get_queue()->get_next( 'oss_woocommerce_' . $id ) ) {
 				return true;
 			}
 		}
@@ -88,11 +87,11 @@ class Queue {
 		return false;
 	}
 
-	public static function is_completed( $type ) {
+	public static function is_completed( $id ) {
 		$running = self::get_reports_running();
 
-		if ( array_key_exists( $type, $running ) && 'completed' === $running[ $type ]['status'] ) {
-			if ( ! self::get_queue()->get_next( 'oss_woocommerce_' . $type ) ) {
+		if ( array_key_exists( $id, $running ) && 'completed' === $running[ $id ] ) {
+			if ( ! self::get_queue()->get_next( 'oss_woocommerce_' . $id ) ) {
 				return true;
 			}
 		}
@@ -118,7 +117,7 @@ class Queue {
 
 			$queue->schedule_single(
 				time() + 10,
-				'oss_woocommerce_' . $type,
+				'oss_woocommerce_' . $generator->get_id(),
 				array( 'args' => $new_args ),
 				'oss_woocommerce'
 			);
@@ -137,42 +136,38 @@ class Queue {
 		/**
 		 * Cancel outstanding events.
 		 */
-		$queue->cancel_all( 'oss_woocommerce_' . $type );
+		$queue->cancel_all( 'oss_woocommerce_' . $generator->get_id() );
 
 		$report = $generator->complete();
 		$status = 'failed';
 
-		if ( is_a( $report, '\Vendidero\OneStopShop\Report' ) ) {
-			$reports_available = Queue::get_report_keys();
+		if ( is_a( $report, '\Vendidero\OneStopShop\Report' ) && $report->exists() ) {
+			$reports_available = Queue::get_report_ids();
 			$status            = 'completed';
 
-			if ( ! in_array( $report->get_id(), $reports_available ) ) {
-				$reports_available[ $type ][] = $report->get_id();
-
+			if ( ! in_array( $report->get_id(), $reports_available[ $type ] ) ) {
+				array_unshift( $reports_available[ $type ], $report->get_id() );
 				update_option( 'oss_woocommerce_reports', $reports_available );
 			}
 		}
 
-		Package::log( sprintf( 'Completed %1$s report. Status: %2$s', self::get_type_title( $type ), $status ) );
+		Package::log( sprintf( 'Completed %1$s. Status: %2$s', $report->get_title(), $status ) );
 
 		$running = self::get_reports_running();
 
-		if ( ! array_key_exists( $type, $running ) ) {
-			$running[ $type ] = array(
-				'status' => $status,
-				'id'     => $generator->get_id()
-			);
+		if ( ! array_key_exists( $generator->get_id(), $running ) ) {
+			$running[ $generator->get_id() ] = $status;
 		} else {
-			$running[ $type ]['status'] = $status;
+			$running[ $generator->get_id() ] = $status;
 		}
 
 		update_option( 'oss_woocommerce_reports_running', $running );
 	}
 
-	protected static function get_report_keys() {
+	public static function get_report_ids() {
 		$reports = (array) get_option( 'oss_woocommerce_reports', array() );
 
-		foreach( array_keys( self::get_available_types() ) as $type ) {
+		foreach( array_keys( Package::get_available_types() ) as $type ) {
 			if ( ! array_key_exists( $type, $reports ) ) {
 				$reports[ $type ] = array();
 			}
@@ -185,13 +180,17 @@ class Queue {
 		return (array) get_option( 'oss_woocommerce_reports_running', array() );
 	}
 
-	public static function get_timeframe( $type, $date = null ) {
+	public static function get_timeframe( $type, $date = null, $date_end = null ) {
 		$date_start      = null;
-		$date_end        = null;
+		$date_end        = is_null( $date_end ) ? null : $date_end;
 		$start_indicator = is_null( $date ) ? new \WC_DateTime() : $date;
 
 		if ( ! is_a( $start_indicator, 'WC_DateTime' ) && is_numeric( $start_indicator ) ) {
 			$start_indicator = new \WC_DateTime( "@" . $start_indicator );
+		}
+
+		if ( ! is_null( $date_end ) && ! is_a( $date_end, 'WC_DateTime' ) && is_numeric( $date_end ) ) {
+			$date_end = new \WC_DateTime( "@" . $date_end );
 		}
 
 		if ( 'quarterly' === $type ) {
@@ -213,16 +212,23 @@ class Queue {
 
 			$date_start = new \WC_DateTime( "first day of " . $start_month . " " . $start_indicator->date_i18n( 'Y' ) . " midnight" );
 			$date_end   = new \WC_DateTime( "last day of " . $end_month . " " . $start_indicator->date_i18n( 'Y' ) . " midnight" );
-		} elseif ( 'monthly' ) {
+		} elseif ( 'monthly' === $type ) {
 			$month = $start_indicator->date_i18n( 'M' );
 
 			$date_start = new \WC_DateTime( "first day of " . $month . " " . $start_indicator->date_i18n( 'Y' ) . " midnight" );
 			$date_end   = new \WC_DateTime( "last day of " . $month . " " . $start_indicator->date_i18n( 'Y' ) . " midnight" );
-		} else {
+		} elseif ( 'yearly' === $type ) {
 			$date_end   = clone $start_indicator;
 			$date_start = clone $date_end;
 
 			$date_start->modify( '-1 year' );
+		} else {
+			if ( is_null( $date_end ) ) {
+				$date_end = clone $start_indicator;
+				$date_end->modify( '-1 year' );
+			}
+
+			$date_start = clone $start_indicator;
 		}
 
 		return array(
